pyproject.toml
[project]
name = "wechatcustomchatbot"
version = "0.1.0"
description = "Single-user, self-hosted WeChat LLM bot brain kit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "APScheduler>=3.10",
  "httpx>=0.27",
  "python-dotenv>=1.0",
]

[project.scripts]
wechat-bot = "bot.main:run"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["bot", "bot.*", "connectors", "connectors.*"]
